Blueprint added, simplified app.py
