+++ index/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+index_bp = Blueprint("index", __name__, url_prefix="/")
+
+from . import routes
