+++ quiz/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+quiz_bp = Blueprint("quiz", __name__, url_prefix="/quiz")
+
+from . import routes
