+++ templates/base.html
@@ -10,27 +10,27 @@
     </head>
     <body>
       <nav class="header">
-        <a href="{{ url_for('home') }}">
+        <a href="{{ url_for('index.home') }}">
           <img src="{{ url_for('static', filename='images/flaskify.jpg') }}" alt="FLASKIFY" class="logo">
         </a>
         
         <ul class="nav-links">
             <li>
-              <form action="{{ url_for('search') }}" method='get'>
+              <form action="{{ url_for('index.search') }}" method='get'>
                   <input class='search' placeholder='Search' type='text' name='query'/>
               </form>
             </li>
-            <li><a href="{{ url_for('show_sets') }}">View Sets</a></li>
-            <li><a href="{{ url_for('create_sets') }}">Create Set</a></li>
-            <li><a href="{{ url_for('create_cards') }}">Create Card</a></li>
-            <li><a href="{{ url_for('select_quiz_sets') }}">Quiz</a></li>
+            <li><a href="{{ url_for('set.show_sets') }}">View Sets</a></li>
+            <li><a href="{{ url_for('set.create_sets') }}">Create Set</a></li>
+            <li><a href="{{ url_for('set.create_cards') }}">Create Card</a></li>
+            <li><a href="{{ url_for('quiz.select_quiz_sets') }}">Quiz</a></li>
             
             <li class="auth-container">
                 {% if session.get('user_id') %}
-                    <a href="{{ url_for('logout') }}">Logout</a>
+                    <a href="{{ url_for('user.logout') }}">Logout</a>
                 {% else %}
-                    <a href="{{ url_for('login') }}">Login </a>
-                    <a href="{{ url_for('signup') }}">Sign Up</a>
+                    <a href="{{ url_for('user.login') }}">Login </a>
+                    <a href="{{ url_for('user.signup') }}">Sign Up</a>
                 {% endif %}
             </li>
         </ul>
