+++ templates/create_cards.html
@@ -7,9 +7,9 @@
 
 {% if no_set %}
     <p class="middle">Create a set before creating your first flashcard!</p>
-    <button class="middle" href="{{ url_for('create_sets') }}">Create Set</button>
+    <button class="middle" href="{{ url_for('set.create_sets') }}">Create Set</button>
 {% else %}
-<form class="middle" action="{{ url_for('create_cards', set_id=set_id) }}" method="POST">
+<form class="middle" action="{{ url_for('set.create_cards', set_id=set_id) }}" method="POST">
     <label for="front">Front (Question):</label>
     <input type="text" id="front" name="front" required><br>
