+++ templates/login.html
@@ -13,5 +13,5 @@
 
     <button class="submit" type="submit">Login</button>
 </form>
-<p class="middle">Don't have an account?<br /> <a class="small" href="{{ url_for('signup') }}">Sign up</a></p>
+<p class="middle">Don't have an account?<br /> <a class="small" href="{{ url_for('user.signup') }}">Sign up</a></p>
 {% endblock %}
