+++ templates/quiz.html
@@ -39,7 +39,7 @@
 {% endif %}
 
 
-<a href="{{ url_for('select_quiz_sets') }}">
+<a href="{{ url_for('quiz.select_quiz_sets') }}">
     <button>Back</button>
 </a>
