+++ templates/select_sets.html
@@ -14,7 +14,7 @@
         </form>
     {% else %}
         <p>There are no sets to start the quiz!</p>
-        <a href="{{ url_for('create_sets') }}"><button>Create a Set</button></a>
+        <a href="{{ url_for('set.create_sets') }}"><button>Create a Set</button></a>
 {% endif %}
 
 {% endblock %}
