+++ templates/sets.html
@@ -4,17 +4,17 @@
 
 {% block content %}
 <h2>Here are the Flashcard sets you have:</h2>
-<a class="btn three" href="{{ url_for('create_sets') }}">Create Set</a>
+<a class="btn three" href="{{ url_for('set.create_sets') }}">Create Set</a>
 <ul>
     {% for set in sets %}
     <div class="set">
         <table class="set">
             <tr><th>Title</th><th>Date</th></tr>
-            <tr><td><a class="link" href="{{ url_for('show_set_cards', set_id=set.id) }}">{{ set.name }} </a></td><td>{{ set.createdate }}</td></tr>
+            <tr><td><a class="link" href="{{ url_for('set.show_set_cards', set_id=set.id) }}">{{ set.name }} </a></td><td>{{ set.createdate }}</td></tr>
         </table>
     </div>
     {% else %}
-        <li>No sets available<a class="link" href="{{ url_for('create_sets') }}">Create one</a></li>
+        <li>No sets available<a class="link" href="{{ url_for('set.create_sets') }}">Create one</a></li>
     {% endfor %}
 </ul>
 {% endblock %}
