+++ templates/signup.html
@@ -16,5 +16,5 @@
 
     <button class="submit" type="submit">Sign Up</button>
 </form>
-<p class="middle">Have an account? <a class="small" href="{{ url_for('login') }}">Login here</a></p>
+<p class="middle">Have an account? <a class="small" href="{{ url_for('user.login') }}">Login here</a></p>
 {% endblock %}
