+++ user/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+user_bp = Blueprint("user", __name__)
+
+from . import routes
